Move keyboard progress flow handling to `MatchSettingsOverlay`
