+++ osu.Game/Screens/OnlinePlay/Multiplayer/Match/BeatmapSelectionControl.cs
@@ -5,15 +5,13 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Bindings;
 using osu.Framework.Screens;
-using osu.Game.Input.Bindings;
 using osu.Game.Online.API;
 using osu.Game.Screens.OnlinePlay.Match.Components;
 
 namespace osu.Game.Screens.OnlinePlay.Multiplayer.Match
 {
-    public class BeatmapSelectionControl : RoomSubScreenComposite, IKeyBindingHandler<GlobalAction>
+    public class BeatmapSelectionControl : RoomSubScreenComposite
     {
         [Resolved]
         private MultiplayerMatchSubScreen matchSubScreen { get; set; }
@@ -74,32 +72,14 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer.Match
             }, true);
         }
 
+        public void BeginSelection() => selectButton.TriggerClick();
+
         private void updateBeatmap()
         {
             if (SelectedItem.Value == null)
                 beatmapPanelContainer.Clear();
             else
                 beatmapPanelContainer.Child = new DrawableRoomPlaylistItem(SelectedItem.Value, false, false);
-        }
-
-        public bool OnPressed(GlobalAction action)
-        {
-            // only handle keyboard input if there is no current selection.
-            if (SelectedItem.Value != null)
-                return false;
-
-            switch (action)
-            {
-                case GlobalAction.Select:
-                    selectButton.TriggerClick();
-                    return true;
-            }
-
-            return false;
-        }
-
-        public void OnReleased(GlobalAction action)
-        {
         }
     }
 }
